Convert crypt class into enigma class
