+++ test/test_helper.rb
@@ -9,7 +9,6 @@ require 'minitest/pride'
 require 'mocha/minitest'
 require './lib/enigma'
 require './lib/crack'
-require './lib/crypt'
 require './lib/key'
 require './lib/offset'
 require './lib/shift'
